# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.10.0
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    tr-engine authenticates **client software, not people**. Every client
    (a dashboard deployment, a script, an upload plugin, a website's
    backend, a phone app) holds its own **API key** and sends it as:

    ```
    Authorization: Bearer tre_<64 hex characters>
    ```

    There are no user accounts, logins, sessions or cookies. How a client
    authenticates its own users, if it has any, is up to the client. The
    operator guide and the client-developer guide are in `docs/auth.md`.

    **Scopes.** A key has one of `listen`, `edit` or `admin`, optionally
    plus `upload`, or only `upload`:

    | Scope | Grants | Implies |
    |---|---|---|
    | `listen` | Read radio data, stream events and audio, mint tickets | |
    | `edit` | Change radio metadata (tags, transcription corrections and reviews, unit-tag suggestions) | `listen` |
    | `admin` | Everything else: keys, anonymous access, identity edits, merges, maintenance, storage, SQL query, pages, imports, console logs, audit log | `edit`, `listen` |
...

...
    what applies to it.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /unit-tags/import:
    post:
      operationId: importUnitTags
      x-scope: admin
      x-restricted: deny
      summary: Upload unit tags CSV
      description: |
        Imports a trunk-recorder unit tags CSV (the `unitTagsFile` format) into
        the units table, for deployments where tr-engine can't read trunk-recorder's
        files via `TR_DIR`. Accepts either `system_id` (must exist) or
        `system_name` (see the parameter).

        CSV format: two columns, `unit_id,alpha_tag`, no header required. A header
        row (first row whose first column is not numeric), a UTF-8 BOM, blank
        lines, quoted fields and extra columns are all accepted. Rows that can't be
        parsed, have a non-numeric or out-of-range unit ID (not 1..2147483647), or have an empty tag
        are counted in `skipped`. If a unit ID repeats, the last row wins.

        Tags are stored with `alpha_tag_source: csv` using the priority
        manual > csv > mqtt: they replace MQTT-discovered tags and previously
        imported CSV tags, and MQTT ingest will not change them afterwards.
        Manually edited units keep their tag. Units in the CSV that have not been
        heard yet are created with no first/last seen time. The file is applied
        in a single statement: either all rows are imported or, on a database
        error (500), none are.
      tags: [units]
      parameters:
        - name: system_id
          in: query
          description: Target system ID (must exist). Mutually exclusive with system_name.
          schema:
            type: integer
        - name: system_name
          in: query
          description: >-
            Target system name. Mutually exclusive with system_id. Selects the
            non-deleted system whose name, or one of whose sites' short name (TR's
            shortName), equals it exactly. The system must already exist: systems are
            created when trunk-recorder first reports them (MQTT, upload, file watch
            or TR_DIR), so an unknown name fails with 404 and nothing is created.
            If several systems match, the request fails with 409 and lists them;
            use system_id instead.
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Unit tags CSV file (unit_id,alpha_tag)
      responses:
        "200":
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                required: [imported, total, system_id, skipped]
                properties:
                  imported:
                    type: integer
                    description: Number of rows imported; always equals total, since the import is all-or-nothing (manually tagged units are counted but keep their tag)
                  total:
                    type: integer
                    description: Total valid rows in the CSV
                  system_id:
                    type: integer
                    description: System ID the unit tags were imported into
                  skipped:
                    type: integer
                    description: Rows skipped as malformed, with an invalid unit ID, or with an empty tag
                  duplicates:
                    type: integer
                    description: Rows repeating a unit ID seen earlier in the file; the last one wins (omitted when 0)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"
  /units:
    get:
      operationId: listUnits
      x-scope: listen
      x-restricted: deny
      summary: List radio units
      description: |
        Returns radio units with optional filters. When searching, results
        include relevance_score (100=exact, 50=prefix, 10=contains).
      tags: [units]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID
          schema:
            type: string
        - name: search
          in: query
          description: Search by alpha_tag or unit_id
          schema:
            type: string
        - name: active_within
          in: query
          description: |
            Only return units active within this many minutes. Replaces the
            need for a separate "active units" endpoint when combined with
            other filters.
          schema:
            type: integer
        - name: talkgroup
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple)
          schema:
            type: string
            example: "9178,5344"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/sortDir"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /units/{id}:
    get:
      operationId: getUnit
      x-scope: listen
      x-restricted: deny
      summary: Get a unit
      description: |
        Returns a single unit. Accepts composite `system_id:unit_id` (e.g.,
        "1:1234567") or plain `unit_id`. Returns 409 if ambiguous.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateUnit
      x-scope: edit
      x-restricted: deny
      summary: Update unit metadata
      description: >-
        Updates mutable unit fields (alpha_tag, alpha_tag_source).
        Setting a non-empty alpha_tag marks the unit `alpha_tag_source: manual`
        (any alpha_tag_source in the same request is ignored), so neither MQTT
        ingest nor a unit tags CSV re-import will overwrite the edit. An empty
        alpha_tag is ignored: it neither clears the tag nor marks it manual. An
        unknown alpha_tag_source value is rejected with 400.
        When CSV_WRITEBACK is enabled and TR_DIR is configured with a unitTagsFile,
        alpha_tag changes are also written back to trunk-recorder's unit CSV file on disk.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"
  /units/{id}/calls:
    get:
      operationId: listUnitCalls
      x-scope: listen
      x-restricted: deny
      summary: List calls for a unit
      description: |
        Returns calls that include transmissions from a specific unit.
        Accepts composite `system_id:unit_id` or plain `unit_id`.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"
  /units/{id}/events:
    get:
      operationId: listUnitEvents
      x-scope: listen
      x-restricted: deny
      summary: List unit events
      description: |
        Returns events (affiliations, registrations, calls, etc.) for a
        specific unit. Includes embedded talkgroup and unit display names.
      tags: [units]
      parameters:
        - $ref: "#/components/parameters/unitId"
        - name: type
          in: query
          description: Filter by event type
          schema:
            $ref: "#/components/schemas/EventType"
        - name: talkgroup
          in: query
          description: Filter by talkgroup ID
          schema:
            type: integer
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitEventListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Ambiguous"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Unit Tag Suggestions (review queue)
  # ----------------------------------------------------------
...
  /calls:
    get:
      operationId: listCalls
      x-scope: listen
      x-restricted: enforced
      summary: List calls
      description: |
        Returns call recordings with comprehensive filters. This is the
        primary endpoint for querying call history.

        **Common patterns:**
        - Live feed: `?sort=-stop_time&deduplicate=true&limit=50`
        - Talkgroup history: `?tgid=9178&sysid=348&start_time=...&end_time=...`
        - Unit activity: `?unit_id=924003`
        - Emergency calls: `?emergency=true&sort=-start_time`
      tags: [calls]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID (comma-separated for multiple, e.g., "348,34D"). Alias "sysids" also accepted.
          schema:
            type: string
        - name: system_id
          in: query
          description: Filter by system database ID (comma-separated for multiple). Alias "systems" also accepted.
          schema:
            type: string
            example: "1,2"
        - name: site_id
          in: query
          description: Filter by recording site ID (comma-separated for multiple). Alias "sites" also accepted.
          schema:
            type: string
            example: "1,3"
        - name: tgid
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple). Alias "tgids" also accepted.
          schema:
            type: string
            example: "9178,5344"
        - name: unit_id
          in: query
          description: Filter by unit radio ID (comma-separated for multiple). Aliases "units", "unit_ids" also accepted.
          schema:
            type: string
            example: "924003,104"
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - name: encrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
        - name: deduplicate
          in: query
          description: |
            When true, returns one call per call_group (best recording).
            Useful for feed views where simulcast duplicates are unwanted.
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - name: sort
          in: query
          description: |
            Sort field. Prefix with `-` for descending.
            Allowed: start_time, stop_time, duration, tgid, freq
          schema:
            type: string
            default: "-start_time"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: tre_<64 hex characters>
      description: |
        An API key in the `Authorization` header:
        `Authorization: Bearer tre_...`. Keys are created by an admin
        (`POST /keys` or `tr-engine keys create`). Keys are never accepted
        from the URL.
    ticketAuth:
      type: apiKey
      in: query
      name: ticket
      description: |
        A short-lived, listen-only ticket from `POST /tickets`, passed as
        `?ticket=trt_...`. Accepted only by `GET`/`HEAD` on
        `/events/stream`, `/audio/live` and `/calls/{id}/audio`, for
        browser APIs that cannot send headers (`EventSource`,
        `<audio src>`, `WebSocket`). Takes precedence over an
        `Authorization` header on those operations.

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
...
    unitId:
      name: id
      in: path
      required: true
      description: "Unit ID: `system_id:unit_id` (e.g., `1:924003`) or plain `unit_id`"
      schema:
        type: string
        example: "1:924003"
...
    startTime:
      name: start_time
      in: query
      description: Start of time range (RFC 3339)
      schema:
        type: string
        format: date-time
...
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
    sort:
      name: sort
      in: query
      description: "Sort field. Prefix with `-` for descending (e.g., `-last_seen`)"
      schema:
        type: string
        default: "alpha_tag"
    sortDir:
      name: sort_dir
      in: query
      description: Sort direction (ignored if sort field uses `-` prefix)
      schema:
        type: string
        enum: [asc, desc]
        default: asc

  # ----------------------------------------------------------
  # Reusable Responses
  # ----------------------------------------------------------
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Ambiguous:
      description: |
        Ambiguous ID — the plain ID (or, for CSV imports, the system_name)
        matches multiple systems. Response includes the list of matching
        systems to help the client disambiguate.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AmbiguousError"
    Unauthorized:
      description: |
        No usable credential. `code` is `key_required` (no credential, and
        the anonymous access policy doesn't allow this operation),
        `invalid_key` (unknown, revoked or expired key) or `invalid_ticket`
        (malformed, expired or orphaned ticket).
      headers:
        WWW-Authenticate:
          description: Always `Bearer realm="tr-engine"`
          schema:
            type: string
            example: Bearer realm="tr-engine"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            key_required:
              value:
                code: key_required
                error: an API key is required
            invalid_key:
              value:
                code: invalid_key
                error: API key has been revoked
    Forbidden:
      description: |
        The credential is valid but not allowed here. `code` is
        `insufficient_scope` (the message names the scope the operation
        needs) or `restricted_credential` (the credential is restricted to
        some systems or talkgroups, and this operation can't enforce that).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficient_scope:
              value:
                code: insufficient_scope
                error: this operation needs the edit scope
            restricted_credential:
              value:
                code: restricted_credential
                error: this operation is not available to restricted credentials
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    Unit:
      type: object
      description: A radio unit (mobile or portable radio)
      required: [system_id, unit_id]
      properties:
        system_id:
          type: integer
          description: Database system ID this unit belongs to
          example: 1
        system_name:
          type: string
          description: System display name (embedded to avoid lookup)
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"
        unit_id:
          type: integer
          example: 1234567
        alpha_tag:
          type: string
          example: Engine 1
        alpha_tag_source:
          type: string
          description: |
            Where the alpha tag came from: `manual` (user edit), `csv` (unit tags
            CSV via TR_DIR or `POST /unit-tags/import`), or `mqtt`. Omitted when
            the tag was discovered from MQTT/audio metadata. Priority is
            manual > csv > mqtt; lower-priority sources never overwrite higher ones.
          example: csv
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        # Last event context (embedded)
        last_event_type:
          $ref: "#/components/schemas/EventType"
        last_event_time:
          type: string
          format: date-time
        last_event_tgid:
          type: integer
          example: 9178
        last_event_tg_tag:
          type: string
          description: Alpha tag of the talkgroup from last event
          example: "09-8L Main"
        # Tag observations (auto-managed, read-only). Stored beside alpha_tag
        # and never change it or its manual > csv > mqtt priority.
        recorder_alpha_tag:
          type: string
          readOnly: true
          description: |
            Latest non-empty unit tag reported by trunk-recorder (`unit_alpha_tag`
            in MQTT unit/call messages, or the srcList `tag` in watch/upload
            modes), recorded even when a manual or CSV `alpha_tag` wins. This is
            trunk-recorder's own resolved tag: its `unitTagsFile` entry if it has
            one, otherwise the decoded over-the-air alias. Omitted if never reported.
          example: "P338 FF1"
        recorder_alpha_tag_seen:
          type: string
          format: date-time
          readOnly: true
          description: When `recorder_alpha_tag` was last reported.
        ota_alpha_tag:
          type: string
          readOnly: true
          description: |
            Latest raw over-the-air unit alias, as reported separately from the
            resolved tag: `unit_alpha_tag_ota` in MQTT unit/call messages (MQTT
            plugin), or `srcList[].tag_ota` in trunk-recorder 5.2+ call JSON
            (watch/upload modes). Latest known value only (not a history);
            replaced when a different alias is reported. trunk-recorder reports
            the latest alias it knows for the unit, so the seen times record when
            tr-engine received that report, not necessarily a fresh decode.
            Omitted until a source that sends it reports this unit.
          example: "P338 FF1"
        ota_alpha_tag_first_seen:
          type: string
          format: date-time
          readOnly: true
          description: |
            When tr-engine first recorded the current `ota_alpha_tag` value.
            Resets when the alias changes. A late report of the same alias with
            an older timestamp does not move it earlier, so it never reaches back
            past an earlier alias change (archive import and system merge move it
            earlier only when the incoming observation window overlaps).
        ota_alpha_tag_last_seen:
          type: string
          format: date-time
          readOnly: true
          description: When the current `ota_alpha_tag` value was last observed.
        call_count:
          type: integer
          description: "Number of calls within the requested time window (only present on /talkgroups/{id}/units)"
          example: 42
        # Search field
        relevance_score:
          type: integer
          description: "Search relevance: 100=exact, 50=prefix, 10=contains. Only present in search results."
          example: 100
...
    UnitListResponse:
      type: object
      required: [units, total, limit, offset]
      properties:
        units:
          type: array
          items:
            $ref: "#/components/schemas/Unit"
        total:
          type: integer
          example: 500
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
...